submissions/vidminal/requirements.txt
Pillow>=9.0.0
pygame
moviepy
py7zr